Time: 5 ms (0%), Space: 122.3 MB (0%) - LeetHub

+++ 3788-maximum-score-of-a-split/3788-maximum-score-of-a-split.java
@@ -18,7 +18,7 @@ class Solution {
         long res=Long.MIN_VALUE;
         for(int i=0;i<nums.length-1;i++)
         {
-            res=(long)Math.max(res,(long)(pref[i]-suff[i+1]));
+            res=Math.max(res,(long)(pref[i]-suff[i+1]));
         }
         return res;
     }
